Consolidate modals and improve test coverage
