+++ src/client/components/modals/CreateRoomModal.jsx
@@ -4,15 +4,16 @@ import Modal    from 'react-modal';
 
 /**
  * Checks if input is valid:
- * - only alphanumeric characters plus single spaces & dashes
+ * - only alphanumeric characters plus dashes
  * @param  {Object} input
  * @return {Boolean}      if input is valid or not
  */
-export const validateInput = (input) => (
-  input && /(^\w+[-]?\w*)[ ]?(\w*[-]?\w+$)/g.test(input)
-);
+export const validateInput = (input) => {
+  const alphanumeric = /^\w+(\w+|[-])\w+$/g.test(input.name);
+  return input.name && input.username && alphanumeric;
+};
 
-export class SignInModal extends React.Component {
+export class CreateRoomModal extends React.Component {
   constructor(props) {
     super(props);
     this.state = {
@@ -23,23 +24,33 @@ export class SignInModal extends React.Component {
     this.handleSubmit = this.handleSubmit.bind(this);
   }
   handleChange(e) {
-    const input = e.target.value;
-    if (!this.state.validInput && validateInput(input)) {
+    const input = {
+      name: e.target.value,
+      username: this.props.user.username
+    };
+    const validInput = validateInput(input);
+    if (!this.state.validInput && validInput) {
+      // if valid input, change input & remove invalid message if any
       this.setState({
-        name: input,
+        name: input.name,
         validInput: true
       });
     } else {
+      // change input but don't show invalid message until submitted
       this.setState({
-        name: input
+        name: input.name
       });
     }
   }
   handleSubmit(e) {
     e.preventDefault();
-    const input = this.state.name;
+    const input = {
+      name: this.state.name,
+      username: this.props.user.username
+    };
     const validInput = validateInput(input);
     if (validInput) {
+      // submit w/ valid input
       this.props.onFormSubmit(input);
       this.setState({
         name: '',
@@ -47,6 +58,7 @@ export class SignInModal extends React.Component {
       });
       this.props.onRequestClose();
     } else {
+      // trigger invalid message on submit
       this.setState({
         validInput: false
       });
@@ -62,7 +74,7 @@ export class SignInModal extends React.Component {
       overlayClassName="modal-overlay"
       >
         <header className="modal-header">
-          <h1 className="modal-title">Sign In</h1>
+          <h1 className="modal-title">Create a room</h1>
           <button
           className="modal-close"
           onClick={this.props.onRequestClose}
@@ -74,20 +86,20 @@ export class SignInModal extends React.Component {
           <input
           onChange={this.handleChange}
           value={this.state.name}
-          placeholder="User Name"
+          placeholder="Room name"
           type="text"
           autoFocus
           autoComplete="off"
           />
           {!this.state.validInput &&
-            <div className="validation-error">Invalid input! Please try Again.</div>}
+            <div className="validation-error">Something went wrong! Sign in and check your input.</div>}
         </section>
         <footer className="modal-footer">
           <button
           className="modal-footer-btn"
           onClick={this.handleSubmit}
           >
-            Sign in
+            Create
           </button>
         </footer>
       </Modal>
@@ -95,10 +107,11 @@ export class SignInModal extends React.Component {
   }
 }
 
-SignInModal.PropTypes = {
+CreateRoomModal.PropTypes = {
   modalIsOpen: PropTypes.bool.isRequired,
   onRequestClose: PropTypes.func.isRequired,
   onFormSubmit: PropTypes.func.isRequired,
+  user: PropTypes.object.isRequired
 };
 
-export default SignInModal;
+export default CreateRoomModal;
