+++ src/client/components/modals/CreateRoomModal.spec.js
@@ -41,11 +41,21 @@ describe('<CreateRoomModal />', () => {
   });
 
   it('should simulate input field', () => {
-    const value = 'My new value';
+    const validInput = 'Webpack';
+    const invalidInput = '*_Webpack_*';
     const { wrapper } = setup();
-    wrapper.find('input').simulate('change', {target: {value}});
+    wrapper.state().validInput = false;
+    expect(wrapper.state().validInput).toBe(false);
+
+    wrapper.find('input').simulate('change', {target: {value: invalidInput}});
     expect(wrapper.find('input').props().value)
-      .toEqual(value);
+    .toEqual(invalidInput);
+    expect(wrapper.state().validInput).toBe(false);
+
+    wrapper.find('input').simulate('change', {target: {value: validInput}});
+    expect(wrapper.find('input').props().value)
+      .toEqual(validInput);
+    expect(wrapper.state().validInput).toBe(true);
   });
 
   it('should call submit action on submit with valid input', () => {
